🔨 Clean up DisplayRenderer

+++ src/commonMain/kotlin/de/tfr/game/lib/actor/Box2D.kt
@@ -1,4 +1,4 @@
 package de.tfr.game.lib.actor
 
 
-class Box2D(point: Point, override var width: Float, override var height: Float) : Point2D(point.x, point.y), Box
+class Box2D(point: Point, override var width: Double, override var height: Double) : Point2D(point.x, point.y), Box
